@@ -14,7 +14,7 @@ from test_cases.utils import (
 """
 iRODS native rules usage summary:
 projectCollection:
-    closeProjectCollection:
+    close_project_collection:
         IN RW, not in MDR
         Used in RS (ingest & tape)
     createProjectCollection
@@ -122,7 +122,7 @@ class TestCollections:
             assert list_collections[collection_index]["creator"] == "jonathan.melius@maastrichtuniversity.nl"
             assert list_collections[collection_index]["numUserFiles"] == 0
             assert list_collections[collection_index]["numFiles"] == 4
-            assert list_collections[collection_index]["size"] == 550514.0
+            assert list_collections[collection_index]["size"] == 544610.0
             collection_id = self.collection_id[:-1] + str(collection_index + 1)
             assert list_collections[collection_index]["id"] == collection_id
             pid_suffix = "{}{}".format(self.project_id, collection_id)
@@ -144,7 +144,7 @@ class TestCollections:
         assert collection_detail["externals"] == "no-externalPID-set"
 
         assert int(collection_detail["numFiles"]) == 4
-        assert int(collection_detail["byteSize"]) == 550514
+        assert int(collection_detail["byteSize"]) == 544610
 
         assert self.manager1 in collection_detail["managers"]["users"]
         assert self.manager2 in collection_detail["managers"]["users"]
@@ -171,19 +171,19 @@ class TestCollections:
         collection_path = "/nlmumc/home/jmelius"
         user_to_check = "auser"
         acl = "ils -A {}".format(collection_path)
-        ret_acl = subprocess.check_output(acl, shell=True)
+        ret_acl = subprocess.check_output(acl, shell=True, encoding="UTF-8")
         assert "{}#nlmumc:own".format(user_to_check) not in ret_acl
 
         rule = '/rules/tests/run_test.sh -r set_acl -a "default,admin:own,{},{}"'.format(user_to_check, collection_path)
         subprocess.check_call(rule, shell=True)
-        ret_acl = subprocess.check_output(acl, shell=True)
+        ret_acl = subprocess.check_output(acl, shell=True, encoding="UTF-8")
         assert "{}#nlmumc:own".format(user_to_check) in ret_acl
 
         rule = '/rules/tests/run_test.sh -r set_acl -a "default,admin:null,{},{}"'.format(
             user_to_check, collection_path
         )
         subprocess.check_call(rule, shell=True)
-        ret_acl = subprocess.check_output(acl, shell=True)
+        ret_acl = subprocess.check_output(acl, shell=True, encoding="UTF-8")
         assert "{}#nlmumc:own".format(user_to_check) not in ret_acl
 
     def test_collection_size_per_resource(self):
@@ -198,25 +198,23 @@ class TestCollections:
             assert list_collections[collection_id][0]["relativeSize"] == 100.0
             assert list_collections[collection_id][0]["resourceId"].isnumeric()
             assert list_collections[collection_id][0]["resourceName"] == self.destination_resource
-            assert list_collections[collection_id][0]["size"] == "550514"
+            assert list_collections[collection_id][0]["size"] == "544610"
 
     def test_project_collection_acl_open_close_state(self):
         project_collection_path = formatters.format_project_collection_path(self.project_id, self.collection_id)
         user_to_check = "rods"
         acl = "ils -A {}".format(project_collection_path)
-        ret_acl = subprocess.check_output(acl, shell=True)
+        ret_acl = subprocess.check_output(acl, shell=True, encoding="UTF-8")
         assert "{}#nlmumc:own".format(user_to_check) not in ret_acl
 
         rule_open = "irule -r irods_rule_engine_plugin-irods_rule_language-instance -F /rules/native_irods_ruleset/projectCollection/openProjectCollection.r \"*project='{}'\" \"*projectCollection='{}'\" \"*user='{}'\" \"*rights='own'\"".format(
             self.project_id, self.collection_id, user_to_check
         )
         subprocess.check_call(rule_open, shell=True)
-        ret_acl = subprocess.check_output(acl, shell=True)
+        ret_acl = subprocess.check_output(acl, shell=True, encoding="UTF-8")
         assert "{}#nlmumc:own".format(user_to_check) in ret_acl
 
-        rule_close = "irule -r irods_rule_engine_plugin-irods_rule_language-instance -F /rules/native_irods_ruleset/projectCollection/closeProjectCollection.r \"*project='{}'\" \"*projectCollection='{}'\" ".format(
-            self.project_id, self.collection_id
-        )
+        rule_close = f'/rules/tests/run_test.sh -r close_project_collection -a "{self.project_id},{self.collection_id}"'
         subprocess.check_call(rule_close, shell=True)
-        ret_acl = subprocess.check_output(acl, shell=True)
+        ret_acl = subprocess.check_output(acl, shell=True, encoding="UTF-8")
         assert "{}#nlmumc:own".format(user_to_check) not in ret_acl
