@@ -43,7 +43,7 @@ def list_contributing_projects_by_attribute(ctx, attribute):
 
     parameters = "COLL_NAME"
     conditions = (
-        "COLL_ACCESS_NAME in ('own', 'modify object') "
+        "COLL_ACCESS_NAME in ('own', 'modify_object') "
         "and COLL_ACCESS_USER_ID in ({}) "
         "and COLL_PARENT_NAME = '/nlmumc/projects' "
         "and META_COLL_ATTR_NAME = '{}' AND META_COLL_ATTR_VALUE = 'true'".format(access_user_id, attribute)
