@@ -78,11 +78,14 @@ class TestPolicies:
         run_iquest = (
             'iquest "%s" "SELECT META_COLL_ATTR_VALUE '
             "WHERE COLL_NAME = '{}' and META_COLL_ATTR_NAME = '{}' \"".format(
-                project["project_path"], ProjectAVUs.LATEST_PROJECT_COLLECTION_NUMBER.value
+                project["project_path"],
+                ProjectAVUs.LATEST_PROJECT_COLLECTION_NUMBER.value,
             )
         )
         current_value = subprocess.check_output(run_iquest, shell=True).strip()
-        collection_path = formatters.format_project_collection_path(project["project_id"], "C000000001")
+        collection_path = formatters.format_project_collection_path(
+            project["project_id"], "C000000001"
+        )
         create_collection = "imkdir {}".format(collection_path)
         subprocess.check_call(create_collection, shell=True)
         new_value = subprocess.check_output(run_iquest, shell=True).strip()
@@ -94,31 +97,49 @@ class TestPolicies:
     def test_post_proc_for_modify_avu_metadata(self):
         """This tests whether toggling the 'enableDropzoneSharing' AVU sets properly the ACLs on the dropzones of the changed project"""
         run_ils = "ils -A /nlmumc/ingest/direct/{}".format(self.token)
-        first_ils_output = subprocess.check_output(run_ils, shell=True)
+        first_ils_output = subprocess.check_output(
+            run_ils, shell=True, encoding="UTF-8"
+        )
         assert self.manager2 in first_ils_output
         set_enable_sharing = "imeta set -C /nlmumc/projects/{project_id} enableDropzoneSharing {{value}}".format(
             project_id=self.project_id
         )
         subprocess.check_call(set_enable_sharing.format(value="false"), shell=True)
-        second_ils_output = subprocess.check_output(run_ils, shell=True)
+        second_ils_output = subprocess.check_output(
+            run_ils, shell=True, encoding="UTF-8"
+        )
         assert self.manager2 not in second_ils_output
         subprocess.check_call(set_enable_sharing.format(value="true"), shell=True)
-        third_ils_output = subprocess.check_output(run_ils, shell=True)
+        third_ils_output = subprocess.check_output(
+            run_ils, shell=True, encoding="UTF-8"
+        )
         assert self.manager2 in third_ils_output
 
     def test_post_proc_for_modify_access_control(self):
         """This tests whether adding a user to a project properly adds the users ACLS to the projects direct dropzones"""
-        change_user_access_to_project = "ichmod -M {{access}} dlinssen /nlmumc/projects/{project_id}".format(
-            project_id=self.project_id
+        change_user_access_to_project = (
+            "ichmod -M {{access}} dlinssen /nlmumc/projects/{project_id}".format(
+                project_id=self.project_id
+            )
         )
         run_ils = "ils -A /nlmumc/ingest/direct/{}".format(self.token)
-        first_ils_output = subprocess.check_output(run_ils, shell=True)
+        first_ils_output = subprocess.check_output(
+            run_ils, shell=True, encoding="UTF-8"
+        )
         assert "dlinssen" not in first_ils_output
-        subprocess.check_call(change_user_access_to_project.format(access="own"), shell=True)
-        second_ils_output = subprocess.check_output(run_ils, shell=True)
+        subprocess.check_call(
+            change_user_access_to_project.format(access="own"), shell=True
+        )
+        second_ils_output = subprocess.check_output(
+            run_ils, shell=True, encoding="UTF-8"
+        )
         assert "dlinssen" in second_ils_output
-        subprocess.check_call(change_user_access_to_project.format(access="null"), shell=True)
-        third_ils_output = subprocess.check_output(run_ils, shell=True)
+        subprocess.check_call(
+            change_user_access_to_project.format(access="null"), shell=True
+        )
+        third_ils_output = subprocess.check_output(
+            run_ils, shell=True, encoding="UTF-8"
+        )
         assert "dlinssen" not in third_ils_output
 
     def test_pep_api_data_obj_put_post(self):
@@ -137,17 +158,21 @@ class TestPolicies:
         subprocess.check_call(put_instance, shell=True)
         # The policy assumes 3 replicas for direct ingest sizeIngested to be triggered (0-stagingresc, 1 and 2).
         # Therefor an extra replica on rootResc is created
-        repl_instance = "irepl -R {} {}/instance.json".format("rootResc", collection_path)
+        repl_instance = "irepl -R {} {}/instance.json".format(
+            "rootResc", collection_path
+        )
         subprocess.check_call(repl_instance, shell=True)
         # Test sizeIngested AVU
         get_size_ingested = "iquest \"%s\" \"SELECT META_COLL_ATTR_VALUE WHERE COLL_NAME = '{}' and META_COLL_ATTR_NAME = 'sizeIngested' \"".format(
             collection_path
         )
-        size_ingested = subprocess.check_output(get_size_ingested, shell=True).rstrip("\n")
+        size_ingested = subprocess.check_output(
+            get_size_ingested, shell=True, encoding="UTF-8"
+        ).rstrip("\n")
         assert int(size_ingested) == 12521
         # Test metadata file ACL
         run_ils = "ils -A /nlmumc/ingest/direct/{}/instance.json".format(self.token)
-        ils_output = subprocess.check_output(run_ils, shell=True)
+        ils_output = subprocess.check_output(run_ils, shell=True, encoding="UTF-8")
         assert "{}#nlmumc:read".format(self.manager1) in ils_output
         assert "{}#nlmumc:own".format(self.manager1) not in ils_output
         # teardown
@@ -159,7 +184,9 @@ class TestPolicies:
         # Setup: Add a non-admin manager to the project
         test_manager = "policy_test_manager"
         create_user(test_manager)
-        mod_acl = "ichmod own {} /nlmumc/projects/{}".format(test_manager, self.project_id)
+        mod_acl = "ichmod own {} /nlmumc/projects/{}".format(
+            test_manager, self.project_id
+        )
         subprocess.check_call(mod_acl, shell=True)
 
         financial_manager = self.manager1
@@ -169,10 +196,19 @@ class TestPolicies:
         # Financial => Only Principal Investigator or Data steward
         financial_avu_to_check = "responsibleCostCenter"
         with pytest.raises(subprocess.CalledProcessError) as e_info:
-            subprocess.check_call(check.format(contributor, self.project_id, financial_avu_to_check), shell=True)
+            subprocess.check_call(
+                check.format(contributor, self.project_id, financial_avu_to_check),
+                shell=True,
+            )
         with pytest.raises(subprocess.CalledProcessError) as e_info:
-            subprocess.check_call(check.format(test_manager, self.project_id, financial_avu_to_check), shell=True)
-        subprocess.check_call(check.format(financial_manager, self.project_id, financial_avu_to_check), shell=True)
+            subprocess.check_call(
+                check.format(test_manager, self.project_id, financial_avu_to_check),
+                shell=True,
+            )
+        subprocess.check_call(
+            check.format(financial_manager, self.project_id, financial_avu_to_check),
+            shell=True,
+        )
 
         # Project settings => only project managers, Principal Investigator or Data steward
         list_project_setting_avu_to_check = [
@@ -185,19 +221,34 @@ class TestPolicies:
         ]
         for avu in list_project_setting_avu_to_check:
             with pytest.raises(subprocess.CalledProcessError) as e_info:
-                subprocess.check_call(check.format(contributor, self.project_id, avu), shell=True)
-            subprocess.check_call(check.format(test_manager, self.project_id, avu), shell=True)
-            subprocess.check_call(check.format(financial_manager, self.project_id, financial_avu_to_check), shell=True)
+                subprocess.check_call(
+                    check.format(contributor, self.project_id, avu), shell=True
+                )
+            subprocess.check_call(
+                check.format(test_manager, self.project_id, avu), shell=True
+            )
+            subprocess.check_call(
+                check.format(
+                    financial_manager, self.project_id, financial_avu_to_check
+                ),
+                shell=True,
+            )
 
         # teardown
         remove_user(test_manager)
 
     def test_pre_proc_for_coll_create_first(self):
         """This tests if a user is allowed to make a dir in a direct dropzone that is already ingesting (they should not be)"""
-        set_dropzone_state = "imeta set -C /nlmumc/ingest/direct/{token} state {{state}}".format(token=self.token)
+        set_dropzone_state = (
+            "imeta -M set -C /nlmumc/ingest/direct/{token} state {{state}}".format(
+                token=self.token
+            )
+        )
         subprocess.check_call(set_dropzone_state.format(state="ingesting"), shell=True)
-        create_coll_when_ingesting = "export clientUserName={} && imkdir /nlmumc/ingest/direct/{}/foobar".format(
-            self.manager1, self.token
+        create_coll_when_ingesting = (
+            "export clientUserName={} && imkdir /nlmumc/ingest/direct/{}/foobar".format(
+                self.manager1, self.token
+            )
         )
         with pytest.raises(subprocess.CalledProcessError) as e_info:
             subprocess.check_call(create_coll_when_ingesting, shell=True)
@@ -213,7 +264,9 @@ class TestPolicies:
 
     def test_pre_proc_for_coll_create_third(self):
         """Check if a regular user is allowed to create a directory in the direct dropzones folder (they should not be)"""
-        direct_dropzone = "export clientUserName=service-pid && imkdir /nlmumc/ingest/direct/foo-bar"
+        direct_dropzone = (
+            "export clientUserName=service-pid && imkdir /nlmumc/ingest/direct/foo-bar"
+        )
         with pytest.raises(subprocess.CalledProcessError) as e_info:
             subprocess.check_call(direct_dropzone, shell=True)
 
@@ -225,17 +278,22 @@ class TestPolicies:
 
     def test_pre_proc_for_data_obj_open(self):
         """Test if an iget of a file that is on tape does not work"""
-        put_file_on_tape = (
-            "export clientUserName={0} && iput -fR arcRescSURF01 {1} /nlmumc/home/{0}/instance.json".format(
-                self.manager1, TMP_INSTANCE_PATH
+        put_file_on_tape = "export clientUserName={0} && iput -fR arcRescSURF01 {1} /nlmumc/home/{0}/instance.json".format(
+            self.manager1, TMP_INSTANCE_PATH
+        )
+        get_file_from_tape = (
+            "export clientUserName={0} && iget /nlmumc/home/{0}/instance.json".format(
+                self.manager1
             )
         )
-        get_file_from_tape = "export clientUserName={0} && iget /nlmumc/home/{0}/instance.json".format(self.manager1)
         subprocess.check_call(put_file_on_tape, shell=True)
         with pytest.raises(subprocess.CalledProcessError) as e_info:
             subprocess.check_call(get_file_from_tape, shell=True)
         subprocess.check_call(
-            "export clientUserName={0} && irm -rf /nlmumc/home/{0}/instance.json".format(self.manager1), shell=True
+            "export clientUserName={0} && irm -rf /nlmumc/home/{0}/instance.json".format(
+                self.manager1
+            ),
+            shell=True,
         )
 
     def test_set_resc_scheme_for_create_first(self):
@@ -249,7 +307,7 @@ class TestPolicies:
         )
         subprocess.check_call(put_instance, shell=True)
         check_resource = "ils -l {}/instance.json".format(collection_path)
-        output = subprocess.check_output(check_resource, shell=True)
+        output = subprocess.check_output(check_resource, shell=True, encoding="UTF-8")
         assert self.destination_resource in output
         # teardown
         subprocess.check_call("irm -rf {}".format(collection_path), shell=True)
@@ -267,46 +325,58 @@ class TestPolicies:
     def test_set_resc_scheme_for_create_third(self):
         """Test if a file put directly in the direct dropzone dir is properly blocked and if files put in a direct dropzone have the correct resource"""
         get_instance()
-        put_instance = (
-            "export clientUserName={user} && iput {instance} /nlmumc/ingest/direct/{{path}}instance_test_3.json".format(
-                user=self.manager1, instance=TMP_INSTANCE_PATH
-            )
+        put_instance = "export clientUserName={user} && iput {instance} /nlmumc/ingest/direct/{{path}}instance_test_3.json".format(
+            user=self.manager1, instance=TMP_INSTANCE_PATH
         )
         with pytest.raises(subprocess.CalledProcessError) as e_info:
             subprocess.check_call(put_instance.format(path=""), shell=True)
         subprocess.check_call(put_instance.format(path=self.token + "/"), shell=True)
         output_ils = subprocess.check_output(
-            "ils -l /nlmumc/ingest/direct/{}/instance_test_3.json".format(self.token), shell=True
+            "ils -l /nlmumc/ingest/direct/{}/instance_test_3.json".format(self.token),
+            shell=True,
+            encoding="UTF-8",
         )
         assert "stagingResc01" in output_ils
-        subprocess.check_call("irm -f /nlmumc/ingest/direct/{}/instance_test_3.json".format(self.token), shell=True)
+        subprocess.check_call(
+            "export clientUserName={user} && irm -f /nlmumc/ingest/direct/{{token}}/instance_test_3.json".format(
+                user=self.manager1, token=self.token
+            ),
+            shell=True,
+        )
 
     def test_set_resc_scheme_for_create_fourth(self):
         """Test if a file put directly in the mounted dropzone dir is properly blocked and if files put in a mounted dropzone have the correct resource"""
         get_instance()
         self.dropzone_type = "mounted"
         token = create_dropzone(self)
-        put_instance = (
-            "export clientUserName={user} && iput {instance} /nlmumc/ingest/zones/{{path}}instance_test_3.json".format(
-                user=self.manager1, instance=TMP_INSTANCE_PATH
-            )
+        put_instance = "export clientUserName={user} && iput {instance} /nlmumc/ingest/zones/{{path}}instance_test_3.json".format(
+            user=self.manager1, instance=TMP_INSTANCE_PATH
         )
         with pytest.raises(subprocess.CalledProcessError) as e_info:
             subprocess.check_call(put_instance.format(path=""), shell=True)
         subprocess.check_call(put_instance.format(path=token + "/"), shell=True)
         output_ils = subprocess.check_output(
-            "ils -l /nlmumc/ingest/zones/{}/instance_test_3.json".format(token), shell=True
+            "ils -l /nlmumc/ingest/zones/{}/instance_test_3.json".format(token),
+            shell=True,
+            encoding="UTF-8",
         )
         assert "stagingResc01" in output_ils
 
         # clean up
-        subprocess.check_call("irm -f /nlmumc/ingest/zones/{}/instance_test_3.json".format(token), shell=True)
+        subprocess.check_call(
+            "irm -f /nlmumc/ingest/zones/{}/instance_test_3.json".format(token),
+            shell=True,
+        )
         remove_dropzone(token, "mounted")
         self.dropzone_type = "direct"
 
     def test_set_resc_scheme_for_create_fifth(self):
         """Test if a file put in a direct dropzone when it is ingesting is properly blocked"""
-        set_dropzone_state = "imeta set -C /nlmumc/ingest/direct/{token} state {{state}}".format(token=self.token)
+        set_dropzone_state = (
+            "imeta -M set -C /nlmumc/ingest/direct/{token} state {{state}}".format(
+                token=self.token
+            )
+        )
         subprocess.check_call(set_dropzone_state.format(state="ingesting"), shell=True)
         put_instance = "export clientUserName={} && iput {} /nlmumc/ingest/direct/{}/instance_test_3.json".format(
             self.manager1, TMP_INSTANCE_PATH, self.token
